Rename to FactRetrievalHistoryCompressionStrategy. Add tests.
