+++ agents/agents-core/src/commonTest/kotlin/ai/koog/agents/core/dsl/extension/AIAgentNodesHistoryCompressionTest.kt
@@ -2,7 +2,6 @@ package ai.koog.agents.core.dsl.extension
 
 import ai.koog.agents.core.agent.AIAgent
 import ai.koog.agents.core.agent.config.AIAgentConfig
-import ai.koog.agents.core.dsl.builder.forwardTo
 import ai.koog.agents.core.dsl.builder.strategy
 import ai.koog.agents.core.tools.ToolRegistry
 import ai.koog.agents.features.eventHandler.feature.EventHandler
@@ -202,6 +201,91 @@ class AIAgentNodesHistoryCompressionTest {
             testExecutor.tldrCount,
             tldrMessages.size,
             "The number of TLDR messages in the final history should match the TLDR count"
+        )
+    }
+
+    @Test
+    fun testNodeLLMCompressHistoryWithFactRetrieval() = runTest {
+        // Create a test LLM executor to track fact-extraction responses
+        val testExecutor = TestLLMExecutor()
+        testExecutor.reset()
+
+        val concepts = listOf(
+            Concept(
+                keyword = "user-preferences",
+                description = "User stated preferences and behaviour",
+                factType = FactType.MULTIPLE
+            ),
+            Concept(
+                keyword = "current-task",
+                description = "The most important task currently being worked on",
+                factType = FactType.SINGLE
+            ),
+        )
+
+        val agentStrategy = strategy<String, String>("test") {
+            val compress by nodeLLMCompressHistory<Unit>(
+                strategy = HistoryCompressionStrategy.FactRetrieval(concepts)
+            )
+
+            edge(nodeStart forwardTo compress transformed { })
+            edge(compress forwardTo nodeFinish transformed { "Done" })
+        }
+
+        val results = mutableListOf<Any?>()
+
+        // Create a prompt with 15 message pairs
+        val agentConfig = AIAgentConfig(
+            prompt = createPromptWithMessages(15),
+            model = OllamaModels.Meta.LLAMA_3_2,
+            maxAgentIterations = 10
+        )
+
+        AIAgent(
+            promptExecutor = testExecutor,
+            strategy = agentStrategy,
+            agentConfig = agentConfig,
+            toolRegistry = ToolRegistry {
+                tool(DummyTool())
+            }
+        ) {
+            install(EventHandler) {
+                onAgentCompleted { eventContext -> results += eventContext.result }
+            }
+        }.use { agent ->
+            agent.run("", null)
+        }
+
+        // After compression, we should have one result
+        assertEquals(1, results.size)
+        assertEquals("Done", results.first())
+
+        // FactRetrieval issues one extraction call per concept
+        assertEquals(
+            concepts.size,
+            testExecutor.factCount,
+            "FactRetrieval strategy should issue one extraction request per configured concept"
+        )
+
+        // No TLDR was needed because facts were extracted successfully
+        assertEquals(
+            0,
+            testExecutor.tldrCount,
+            "FactRetrieval strategy should not fall back to TLDR when facts are extracted successfully"
+        )
+
+        // The fact-extraction system prompts should reference all configured concept keywords.
+        val systemMessages = testExecutor.messages.filterIsInstance<Message.System>()
+        assertTrue(
+            concepts.all { c -> systemMessages.any { it.content.contains(c.keyword) } },
+            "Each configured concept keyword must appear in the fact-extraction system prompts"
+        )
+        // Each extraction request must wrap the conversation history in the dedicated XML tag.
+        val userMessages = testExecutor.messages.filterIsInstance<Message.User>()
+        assertEquals(
+            concepts.size,
+            userMessages.count { it.content.contains("<conversation_to_extract_facts>") },
+            "Each concept must be extracted from a conversation wrapped in <conversation_to_extract_facts>"
         )
     }
 }
