+++ agents/agents-core/src/commonTest/kotlin/ai/koog/agents/core/dsl/extension/TestLLMExecutor.kt
@@ -25,12 +25,17 @@ class TestLLMExecutor : PromptExecutor() {
     var tldrCount = 0
         private set
 
+    // Track the number of fact-extraction responses produced
+    var factCount = 0
+        private set
+
     // Store the messages for inspection
     val messages = mutableListOf<Message>()
 
     // Reset the state for a new test
     fun reset() {
         tldrCount = 0
+        factCount = 0
         messages.clear()
     }
 
@@ -68,6 +73,27 @@ class TestLLMExecutor : PromptExecutor() {
             )
             messages.add(tldrResponse)
             return tldrResponse
+        }
+
+        // For FactRetrieval compression: return a structured JSON response
+        if (prompt.messages.any { it.content.contains("specialized information extractor") }) {
+            factCount++
+            val isMultiple = prompt.messages.any {
+                it.content.contains("\"facts\" array") || it.content.contains("facts found")
+            }
+            val factResponse = if (isMultiple) {
+                Message.Assistant(
+                    """{"facts": [{"fact": "Extracted fact #$factCount A"}, {"fact": "Extracted fact #$factCount B"}]}""",
+                    metaInfo = ResponseMetaInfo.create(testClock)
+                )
+            } else {
+                Message.Assistant(
+                    """{"fact": "Extracted single fact #$factCount"}""",
+                    metaInfo = ResponseMetaInfo.create(testClock)
+                )
+            }
+            messages.add(factResponse)
+            return factResponse
         }
 
         val response = Message.Assistant(DEFAULT_ASSISTANT_RESPONSE, metaInfo = ResponseMetaInfo.create(testClock))
