+++ examples/devoxx-belgium-2025/src/main/kotlin/ai/koog/spring/sandwich/agents/AgentStrategy.kt
@@ -11,7 +11,7 @@ import ai.koog.agents.ext.agent.subgraphWithVerification
 import ai.koog.agents.memory.feature.history.RetrieveFactsFromHistory
 import ai.koog.agents.memory.model.Concept
 import ai.koog.agents.memory.model.FactType
-//import ai.koog.agents.core.dsl.extension.ConceptBasedHistoryCompressionStrategy
+//import ai.koog.agents.core.dsl.extension.FactRetrievalHistoryCompressionStrategy
 //import ai.koog.agents.core.dsl.extension.Concept
 //import ai.koog.agents.core.dsl.extension.FactType
 import ai.koog.prompt.executor.clients.anthropic.AnthropicModels
