+++ examples/javaone-2026/src/main/java/org/example/koog/java/agents/KoogAgentService.java
@@ -18,7 +18,7 @@ import ai.koog.agents.features.persistence.jdbc.PostgresJdbcPersistenceStoragePr
 import ai.koog.agents.memory.feature.history.RetrieveFactsFromHistory;
 import ai.koog.agents.memory.model.Concept;
 import ai.koog.agents.memory.model.FactType;
-//import ai.koog.agents.core.dsl.extension.ConceptBasedHistoryCompressionStrategy;
+//import ai.koog.agents.core.dsl.extension.FactRetrievalHistoryCompressionStrategy;
 //import ai.koog.agents.core.dsl.extension.Concept;
 //import ai.koog.agents.core.dsl.extension.FactType;
 import ai.koog.agents.snapshot.feature.Persistence;
