+++ examples/spring-io-2026/java/src/main/java/org/example/koog/spring/java/agents/BankingAgentService.java
@@ -19,7 +19,7 @@ import ai.koog.agents.longtermmemory.retrieval.augmentation.UserPromptAugmenter;
 import ai.koog.agents.memory.feature.history.RetrieveFactsFromHistory;
 import ai.koog.agents.memory.model.Concept;
 import ai.koog.agents.memory.model.FactType;
-//import ai.koog.agents.core.dsl.extension.ConceptBasedHistoryCompressionStrategy;
+//import ai.koog.agents.core.dsl.extension.FactRetrievalHistoryCompressionStrategy;
 //import ai.koog.agents.core.dsl.extension.Concept;
 //import ai.koog.agents.core.dsl.extension.FactType;
 import ai.koog.agents.snapshot.feature.Persistence;
